pyproject.toml
[tool.poetry]
name = "ml-playground"
version = "0.0.1"
description = "ML Playground"
authors = ["WAP"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
audio-recorder-streamlit = "^0.0.8"
chromadb = "^0.5.0"
faster-whisper = "^1.0.1"
ffmpeg-python = "^0.2.0"
langchain = "^0.3.2"
langchain-chroma = "^0.1.1"
langchain-community = "^0.3.1"
langchain-huggingface = "^0.1.0"
langchain-ollama = "^0.2.0"
llama-index = "^0.11.16"
llama-index-embeddings-huggingface = "^0.3.1"
llama-index-embeddings-ollama = "^0.3.1"
llama-index-llms-ollama = "^0.3.3"
llama-index-readers-web = "^0.2.3"
llama-index-vector-stores-chroma = "^0.2.0"
pypdf = "^4.2.0"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.1.1"
streamlit = "^1.32.2"
torch = "^2.3.0"
unstructured = {extras = ["all-docs"], version = "^0.15.14"}
watchdog = "^5.0.3"

[tool.poetry.group.dev.dependencies]
deptry = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
